Extract shared Orchestrator UI to reduce code duplication.
Move repeated table filters, info card layout, and permission batch logic into reusable modules to address Sonar duplication findings.
