+++ workspaces/orchestrator/plugins/orchestrator/src/components/ui/FullWidthInfoCard.tsx
@@ -1,0 +1,35 @@
+/*
+ * Copyright Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { PropsWithChildren, ReactNode } from 'react';
+
+import { InfoCard } from '@backstage/core-components';
+
+import Box from '@mui/material/Box';
+
+export const FullWidthInfoCard = ({
+  title,
+  children,
+}: PropsWithChildren<{ title: ReactNode }>) => (
+  <Box sx={{ width: '100%', maxWidth: '100%', minWidth: 0 }}>
+    <InfoCard
+      title={title}
+      titleTypographyProps={{ component: 'div', style: { width: '100%' } }}
+    >
+      {children}
+    </InfoCard>
+  </Box>
+);
