+++ workspaces/orchestrator/plugins/orchestrator/src/components/ui/TableTextFilter.tsx
@@ -1,0 +1,57 @@
+/*
+ * Copyright Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import ClearIcon from '@mui/icons-material/Clear';
+import SearchIcon from '@mui/icons-material/Search';
+import IconButton from '@mui/material/IconButton';
+import InputAdornment from '@mui/material/InputAdornment';
+import TextField from '@mui/material/TextField';
+
+export const TableTextFilter = ({
+  value,
+  onChange,
+}: {
+  value: string;
+  onChange: (value: string) => void;
+}) => (
+  <TextField
+    variant="standard"
+    size="small"
+    placeholder="Filter"
+    value={value}
+    onChange={event => onChange(event.target.value)}
+    inputProps={{ 'aria-label': 'Filter' }}
+    sx={{ minWidth: { xs: '12rem', sm: '16rem' } }}
+    InputProps={{
+      startAdornment: (
+        <InputAdornment position="start">
+          <SearchIcon fontSize="small" />
+        </InputAdornment>
+      ),
+      endAdornment: value ? (
+        <InputAdornment position="end">
+          <IconButton
+            size="small"
+            aria-label="Clear all"
+            onClick={() => onChange('')}
+          >
+            <ClearIcon fontSize="small" />
+          </IconButton>
+        </InputAdornment>
+      ) : null,
+    }}
+  />
+);
