+++ workspaces/orchestrator/plugins/orchestrator/src/hooks/useWorkflowPermissionBatch.ts
@@ -1,0 +1,45 @@
+/*
+ * Copyright Red Hat, Inc.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { BasicPermission } from '@backstage/plugin-permission-common';
+import { usePermission } from '@backstage/plugin-permission-react';
+
+import { WorkflowOverviewDTO } from '@red-hat-developer-hub/backstage-plugin-orchestrator-common';
+
+import { usePermissionArray } from './usePermissionArray';
+
+export const useWorkflowPermissionBatch = (
+  items: WorkflowOverviewDTO[],
+  genericPermission: BasicPermission,
+  specificPermissionFactory: (workflowId: string) => BasicPermission,
+): { allowed: boolean[] } => {
+  const generic = usePermission({
+    permission: genericPermission,
+  });
+
+  let workflowIds: string[] = [];
+  if (!generic.loading && !generic.allowed) {
+    workflowIds = items.map(item => item.workflowId);
+  }
+
+  const specific = usePermissionArray(
+    workflowIds.map(workflowId => specificPermissionFactory(workflowId)),
+  );
+
+  return {
+    allowed: items.map((_, idx) => generic.allowed || specific.allowed[idx]),
+  };
+};
